Handle org resources specially (#53)
We cannot handle the access control of orgs and projects with
the current mechanism, which only works when a resource is project-scoped.

For example, when a user attempts to list all projects in a given org,
we need to let a caller know which project is visible. On the other hand,
the current authorization API checks if a user is allowed to read/write to
a resource in a specific project and org.

We could extend the API, but it is simpler to implement the auth logic
in the caller (= user-manager).