Bump user-manager dep to llmariner
Once this is deployed, we can remove the legacy API from user-manager.

We don't necessarily need to take this procedure for internal services
that are closed in the control plane, but I wanted to exercise the flow
as we will take the same process for the communication between
the control-plane and the worker-plane.